Refactor secrets and tokens (#18812)
