+++ projects/plugins/jetpack/extensions/blocks/premium-content/_inc/subscription-service/class-jetpack-token-subscription-service.php
@@ -8,7 +8,7 @@
 
 namespace Automattic\Jetpack\Extensions\Premium_Content\Subscription_Service;
 
-use Automattic\Jetpack\Connection\Manager;
+use Automattic\Jetpack\Connection\Tokens;
 
 /**
  * Class Jetpack_Token_Subscription_Service
@@ -41,8 +41,7 @@ class Jetpack_Token_Subscription_Service extends Token_Subscription_Service {
 	 * @return string The key.
 	 */
 	public function get_key() {
-		$connection = new Manager();
-		$token      = $connection->get_access_token();
+		$token = ( new Tokens() )->get_access_token();
 		if ( ! isset( $token->secret ) ) {
 			return false;
 		}
