+++ projects/plugins/jetpack/load-jetpack.php
@@ -40,7 +40,6 @@ add_filter( 'jetpack_should_use_minified_assets', 'jetpack_should_use_minified_a
 // @todo: Abstract out the admin functions, and only include them if is_admin()
 require_once JETPACK__PLUGIN_DIR . 'class.jetpack.php';
 require_once JETPACK__PLUGIN_DIR . 'class.jetpack-network.php';
-require_once JETPACK__PLUGIN_DIR . 'class.jetpack-data.php';
 require_once JETPACK__PLUGIN_DIR . 'class.jetpack-client-server.php';
 require_once JETPACK__PLUGIN_DIR . 'class.jetpack-user-agent.php';
 require_once JETPACK__PLUGIN_DIR . 'class.jetpack-post-images.php';
