+++ projects/plugins/jetpack/tests/php/general/test_class.jetpack-client-server.php
@@ -1,5 +1,4 @@
 <?php
-
 class WP_Test_Jetpack_Client_Server extends WP_UnitTestCase {
 
 	static public function setUpBeforeClass() {
