+++ projects/plugins/jetpack/tests/php/general/test_class.jetpack-xmlrpc-server.php
@@ -1,7 +1,6 @@
 <?php
 
-use Automattic\Jetpack\Sync\Sender;
-use Automattic\Jetpack\Connection\Utils as Connection_Utils;
+use Automattic\Jetpack\Connection\Tokens;
 
 class WP_Test_Jetpack_XMLRPC_Server extends WP_UnitTestCase {
 	static $xmlrpc_admin = 0;
@@ -10,7 +9,7 @@ class WP_Test_Jetpack_XMLRPC_Server extends WP_UnitTestCase {
 		$user_id = $factory->user->create();
 		$user = get_user_by( 'ID', $user_id );
 		$user->set_role( 'administrator' );
-		Connection_Utils::update_user_token( $user_id, sprintf( '%s.%s.%d', 'key', 'private', $user_id ), false );
+		( new Tokens() )->update_user_token( $user_id, sprintf( '%s.%s.%d', 'key', 'private', $user_id ), false );
 
 		self::$xmlrpc_admin = $user_id;
 	}
@@ -180,7 +179,7 @@ class WP_Test_Jetpack_XMLRPC_Server extends WP_UnitTestCase {
 			'blog_token'  => 1,
 			'id'          => 1001,
 		) );
-		Connection_Utils::update_user_token( 1, sprintf( '%s.%d', 'token', 1 ), true );
+		( new Tokens() )->update_user_token( 1, sprintf( '%s.%d', 'token', 1 ), true );
 
 		$server = new Jetpack_XMLRPC_Server();
