+++ tools/phpcs-excludelist.json
@@ -67,7 +67,6 @@
 	"projects/plugins/jetpack/class.jetpack-cli.php",
 	"projects/plugins/jetpack/class.jetpack-client-server.php",
 	"projects/plugins/jetpack/class.jetpack-connection-banner.php",
-	"projects/plugins/jetpack/class.jetpack-data.php",
 	"projects/plugins/jetpack/class.jetpack-heartbeat.php",
 	"projects/plugins/jetpack/class.jetpack-idc.php",
 	"projects/plugins/jetpack/class.jetpack-modules-list-table.php",
@@ -431,7 +430,6 @@
 	"projects/plugins/jetpack/tests/php/general/test-class.jetpack-gutenberg.php",
 	"projects/plugins/jetpack/tests/php/general/test-class.jetpack-options.php",
 	"projects/plugins/jetpack/tests/php/general/test_class.jetpack-client-server.php",
-	"projects/plugins/jetpack/tests/php/general/test_class.jetpack-data.php",
 	"projects/plugins/jetpack/tests/php/general/test_class.jetpack-heartbeat.php",
 	"projects/plugins/jetpack/tests/php/general/test_class.jetpack-plan.php",
 	"projects/plugins/jetpack/tests/php/general/test_class.jetpack-xmlrpc-server.php",
